rename reducers/reducers to reducers/index
