+++ src/app/shared/filter/episodes.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Subscription } from 'rxjs/subscription';
-import { selectEpisodeFilter, selectEpisodes } from '../../ngrx/reducers/reducers';
+import { selectEpisodeFilter, selectEpisodes } from '../../ngrx/reducers';
 import { EpisodeModel } from '../../ngrx/model';
 import { CastleFilterAction } from '../../ngrx/actions';
